metrics_api/requirements.txt
fastapi
uvicorn[standard]
watchdog
PyYAML
python-bitcoinrpc
httpx>=0.24.0
pytz